chore(geo-layers): WMSLayer API audit
