--- website/src/examples/imagery-layer.mdx
@@ -1,5 +1,0 @@
-# ImageryLayer
-
-import Demo from './imagery-layer';
-
-<Demo />
