+++ website/src/examples/wms-layer.mdx
@@ -1,0 +1,5 @@
+# WMSLayer
+
+import Demo from './wms-layer';
+
+<Demo />
